requirements.txt
# AI Knowledge Filler - Python Dependencies
pyyaml>=6.0
anthropic>=0.18.0
python-dotenv>=1.0.0
mypy>=1.8.0
black>=24.0.0
black>=24.0.0
pydocstyle>=6.3.0
